Refactor: split views, add fetch cache and unified error reporting
## SearchView 分割 (459 → 317 行)

- `SearchFilterBars` を抽出: 状態フィルタバー + メモ/コメント+カラーバーの 2 行
- 共有 enum `SearchContentMode` を SearchFilterBars.swift に定義

## EditEntryView 分割 (519 → 479 行)

- `EditEntryStatusSection` を抽出: 種類 (連載/読み切り) + 掲載状況 + 読書状況
  の 3 セクション。読み切り切替時の invariant 矯正もこちらに集約

## N+1 fetch のキャッシュ層

`MangaViewModel.allEntries / allComments / allActivities` に
refreshCounter ベースの簡易キャッシュを追加:

- 同一 render 内で複数 View から呼ばれても 1 回の fetch で済む
- refreshCounter が変わるとキャッシュは自動破棄され次回 fetch が走る
- @ObservationIgnored で View の依存追跡から外し、実際の依存は
  refreshCounter 経由で確保

## 統一エラーハンドリング

新規 `AppError` 型を追加して移行/インポート/エクスポート/保存エラーを
共通フォーマットに集約:

- `MangaViewModel.lastError: AppError?` を導入
- `migrateLegacyStateIfNeeded` の失敗時に lastError をセット (従来は print のみ)
- `RootTabView` に lastError 監視 alert を追加し、どのタブで発生しても
  ユーザーに通知される
