+++ MangaLauncher/Views/Search/SearchView.swift
@@ -21,12 +21,7 @@ struct SearchView: View {
     @State private var commentingEntry: MangaEntry?
     @AppStorage(UserDefaultsKeys.browserMode) private var browserMode: String = "external"
 
-    private let colorLabelStore = ColorLabelStore.shared
     private var theme: ThemeStyle { ThemeManager.shared.style }
-
-    enum SearchContentMode {
-        case entries, memo, comment
-    }
 
     // MARK: - Search Results
 
@@ -133,8 +128,13 @@ struct SearchView: View {
                 }
 
                 VStack(spacing: 0) {
-                    stateFilterBar
-                    secondaryFilterBar
+                    SearchFilterBars(
+                        publicationFilter: $publicationFilter,
+                        readingFilter: $readingFilter,
+                        showOneShotOnly: $showOneShotOnly,
+                        contentMode: $contentMode,
+                        selectedColors: $selectedColors
+                    )
                     content
                 }
             }
@@ -190,148 +190,6 @@ struct SearchView: View {
         .onMangaDataChange {
             viewModel.refresh()
         }
-    }
-
-    // MARK: - 行 1: 状態フィルタ（2 軸）
-
-    @ViewBuilder
-    private var stateFilterBar: some View {
-        ScrollView(.horizontal, showsIndicators: false) {
-            HStack(spacing: 6) {
-                // 掲載状況 + 読み切り（排他グループ）
-                ForEach(PublicationStatus.allCases) { status in
-                    filterChip(label: status.displayName, isSelected: !showOneShotOnly && publicationFilter == status) {
-                        withAnimation(.easeInOut(duration: 0.2)) {
-                            contentMode = .entries
-                            showOneShotOnly = false
-                            publicationFilter = publicationFilter == status ? nil : status
-                        }
-                    }
-                }
-                filterChip(label: "読み切り", isSelected: showOneShotOnly) {
-                    withAnimation(.easeInOut(duration: 0.2)) {
-                        contentMode = .entries
-                        if showOneShotOnly {
-                            showOneShotOnly = false
-                        } else {
-                            showOneShotOnly = true
-                            publicationFilter = nil  // 掲載状況をクリア
-                        }
-                    }
-                }
-
-                Spacer().frame(width: 12)
-
-                // 読書状況
-                ForEach(ReadingState.allCases) { state in
-                    filterChip(label: state.displayName, isSelected: readingFilter == state) {
-                        withAnimation(.easeInOut(duration: 0.2)) {
-                            contentMode = .entries
-                            readingFilter = readingFilter == state ? nil : state
-                        }
-                    }
-                }
-            }
-            .padding(.horizontal)
-            .padding(.vertical, 8)
-        }
-    }
-
-    // MARK: - 行 2: 種別 + メモ/コメント + カラーフィルタ
-
-    @ViewBuilder
-    private var secondaryFilterBar: some View {
-        ScrollView(.horizontal, showsIndicators: false) {
-            HStack(spacing: 8) {
-                // メモ/コメントモード
-                filterChip(label: "メモ", isSelected: contentMode == .memo) {
-                    withAnimation(.easeInOut(duration: 0.2)) {
-                        if contentMode == .memo {
-                            contentMode = .entries
-                        } else {
-                            contentMode = .memo
-                            clearEntryFilters()
-                        }
-                    }
-                }
-                filterChip(label: "コメント", isSelected: contentMode == .comment) {
-                    withAnimation(.easeInOut(duration: 0.2)) {
-                        if contentMode == .comment {
-                            contentMode = .entries
-                        } else {
-                            contentMode = .comment
-                            clearEntryFilters()
-                        }
-                    }
-                }
-
-                // カラーフィルタ
-                ForEach(MangaColor.all) { mangaColor in
-                    let isSelected = selectedColors.contains(mangaColor.name)
-                    let label = colorLabelStore.label(for: mangaColor.name)
-                    Button {
-                        withAnimation(.easeInOut(duration: 0.2)) {
-                            if isSelected {
-                                selectedColors.remove(mangaColor.name)
-                            } else {
-                                selectedColors.insert(mangaColor.name)
-                            }
-                        }
-                    } label: {
-                        HStack(spacing: 6) {
-                            Circle()
-                                .fill(mangaColor.color)
-                                .frame(width: 14, height: 14)
-                            if let label {
-                                Text(label)
-                                    .font(theme.captionFont)
-                            }
-                        }
-                        .padding(.horizontal, 12)
-                        .padding(.vertical, 6)
-                        .background(
-                            isSelected
-                                ? AnyShapeStyle(theme.primary.opacity(0.2))
-                                : AnyShapeStyle(theme.surfaceContainerHigh)
-                        )
-                        .overlay(
-                            Capsule()
-                                .strokeBorder(isSelected ? theme.primary : Color.clear, lineWidth: 1.5)
-                        )
-                        .foregroundStyle(theme.onSurface)
-                        .clipShape(Capsule())
-                    }
-                    .buttonStyle(.plain)
-                }
-            }
-            .padding(.horizontal)
-            .padding(.bottom, 8)
-        }
-    }
-
-    // MARK: - Helpers
-
-    private func filterChip(label: String, isSelected: Bool, action: @escaping () -> Void) -> some View {
-        Button(action: action) {
-            Text(label)
-                .font(theme.bodyFont)
-                .padding(.horizontal, 14)
-                .padding(.vertical, 7)
-                .background(
-                    isSelected
-                        ? AnyShapeStyle(theme.primary)
-                        : AnyShapeStyle(theme.surfaceContainerHigh)
-                )
-                .foregroundStyle(isSelected ? theme.onPrimary : theme.onSurface)
-                .clipShape(Capsule())
-        }
-        .buttonStyle(.plain)
-    }
-
-    private func clearEntryFilters() {
-        publicationFilter = nil
-        readingFilter = nil
-        showOneShotOnly = false
     }
 
     // MARK: - Content
